expand - fix missing var expansion
